conference-registration-app/backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-RESTful
python-dotenv